+++ NEWS.md
@@ -4,6 +4,12 @@ Changes in this release:
 
 * Removed deprecated `run_sea()`. Be sure to use `sea()` now.
 
+* Removed deprecated `get_ggplot()`, please use `plot_demograph()` now.
+
+* `composite()` now returns an empty fhx object if no composite-worthy events are found (Issue #131). Much better than throwing an obtuse error.
+
+* Minor internal code cleanup (Issue #130, Issue #88).
+
 
 # burnr v0.4.0